requirements.txt
PyYAML==6.0.2
python-docx==1.2.0
fonttools==4.59.2
sentence-transformers==5.1.0
